+++ Controls/propertyGrid.ts
@@ -29,7 +29,6 @@
  */
 
 import {default as PropertyGrid} from 'Controls/_propertyGrid/PropertyGrid';
-import {default as View} from 'Controls/_propertyGrid/View';
 import BooleanEditor = require("Controls/_propertyGrid/defaultEditors/Boolean");
 import StringEditor = require("Controls/_propertyGrid/defaultEditors/String");
 import TextEditor = require("Controls/_propertyGrid/defaultEditors/Text");
@@ -40,7 +39,6 @@ import IProperty = require("Controls/_propertyGrid/IProperty");
 import GroupTemplate = require("wml!Controls/_propertyGrid/Render/resources/groupTemplate");
 
 export {
-    View,
     PropertyGrid,
     BooleanEditor,
     StringEditor,
